+++ FireAxe.GUI/ViewModels/AddonNodeSortMethod.cs
@@ -7,7 +7,10 @@ public enum AddonNodeSortMethod
     None,
     Name,
     EnableState,
+    Priority,
+    PriorityInHierarchy,
+    ProblemCount,
+    Tag,
     CreationTime,
     FileSize,
-    Tag
 }